// web/src/api-generated/types.gen.ts
export type AddTodoResponse = {
    /**
     * Id
     */
    id: string;
    /**
     * User Id
     */
    user_id: string;
    /**
     * Title
     */
    title: string;
    /**
     * Is Completed
     */
    is_completed?: boolean;
};